cron job

+++ app/model/channel.py
@@ -1,9 +1,10 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 from .base import Base
+
 
 class Channel(Base):
     __tablename__ = 'channel'
 
-    id = Column(Integer, primary_key=True, autoincrement=True) 
+    id = Column(Integer, primary_key=True, autoincrement=True)
     username = Column(String(128), nullable=False)
-    last_id = Column(Integer, server_default=None, nullable=True)
+    last_id = Column(BigInteger, server_default=None, nullable=True)
